open3d/requirements_depth_map.txt
# create_depth_map
torch 
torchvision
opencv-python 
pillow
rembg 
onnxruntime